+++ src/client/api.ts
@@ -17,7 +17,7 @@ export class StorageApi {
 
   async get({ key, sig, ttl }: { key: string; sig: string; ttl?: TTL }) {
     const params: GetValueParams = prepareQueryParams({ key, sig, ttl });
-    const res = await this.http.get('/get', params);
+    const res = await this.http.get('get', params);
 
     if (res.status !== 200 && res.status !== 404) {
       throw await buildHttpError(res, `Failed to get key "${key}":`);
@@ -28,14 +28,14 @@ export class StorageApi {
 
   async set({ key, value, error }: { key: string; value: unknown; error?: Error }) {
     const body: SetValueParams = { key, value, error: error?.message };
-    const res = await this.http.post('/set', body);
+    const res = await this.http.post('set', body);
     await throwIfHttpError(res, `Failed to save key "${key}":`);
 
     return (await res.json()) as ValueInfo;
   }
 
   async getStale({ key }: GetStaleParams) {
-    const res = await this.http.get('/get-stale', { key });
+    const res = await this.http.get('get-stale', { key });
     await throwIfHttpError(res, `Failed to get stale key "${key}":`);
     const valueInfo = (await res.json()) as ValueInfo | null;
 
@@ -43,7 +43,7 @@ export class StorageApi {
   }
 
   async getStaleList({ prefix }: GetStaleListParams) {
-    const res = await this.http.get('/get-stale-list', { prefix });
+    const res = await this.http.get('get-stale-list', { prefix });
     await throwIfHttpError(res, `Failed to get stale list for prefix "${prefix}":`);
     const valueInfoList = (await res.json()) as ValueInfo[];
 
@@ -51,7 +51,7 @@ export class StorageApi {
   }
 
   async clearSession() {
-    const res = await this.http.post('/clear-session');
+    const res = await this.http.post('clear-session');
     await throwIfHttpError(res, 'Failed to clear session:');
   }
 }