+++ src/server/routes/get-stale-list.ts
@@ -8,11 +8,12 @@ export type GetStaleListParams = {
   prefix: string;
 };
 
-router.get('/get-stale-list', async (req, res) => {
+router.get('/run/:runId/get-stale-list', async (req, res) => {
+  const { runId } = req.params;
   const { prefix } = req.query as GetStaleListParams;
   const config = getConfig(req.app as Express);
 
-  const storage = getStorage(config);
+  const storage = getStorage(config, runId);
   const valueInfoList = await storage.getLoadedInfoList(prefix);
 
   res.json(valueInfoList);